Some namespace and folder re-structuring
