+++ include/cvGPUSpeedup.h
@@ -14,8 +14,8 @@
 
 #pragma once
 
-#include "fast_kernel.h"
 #include "cvGPUSpeedupHelpers.h"
+#include <fast_kernel/fast_kernel.h>
 
 #include <opencv2/core.hpp>
 #include <opencv2/core/cuda_stream_accessor.hpp>
@@ -26,7 +26,7 @@
 namespace cvGS {
 
 template <int I, int O>
-unary_operation_scalar<unary_cuda_vector_cast<CUDA_T(I), CUDA_T(O)>, CUDA_T(I), CUDA_T(O)> convertTo() {
+fk::unary_operation_scalar<fk::unary_cuda_vector_cast<CUDA_T(I), CUDA_T(O)>, CUDA_T(I), CUDA_T(O)> convertTo() {
     return {};
 }
 
@@ -34,28 +34,28 @@ unary_operation_scalar<unary_cuda_vector_cast<CUDA_T(I), CUDA_T(O)>, CUDA_T(I), 
 // They only work with types that have 3 components. In the future
 // they will work with anything.
 template <int I>
-binary_operation_scalar<binary_mul<CUDA_T(I), CUDA_T(I)>, CUDA_T(I), CUDA_T(I)> multiply(cv::Scalar src2) {
-    return operate_t<I, binary_operation_scalar<binary_mul<CUDA_T(I), CUDA_T(I)>, CUDA_T(I), CUDA_T(I)>>()(src2);
+fk::binary_operation_scalar<fk::binary_mul<CUDA_T(I), CUDA_T(I)>, CUDA_T(I), CUDA_T(I)> multiply(cv::Scalar src2) {
+    return internal::operate_t<I, fk::binary_operation_scalar<fk::binary_mul<CUDA_T(I), CUDA_T(I)>, CUDA_T(I), CUDA_T(I)>>()(src2);
 }
 
 template <int I>
-binary_operation_scalar<binary_sub<CUDA_T(I)>, CUDA_T(I), CUDA_T(I)> subtract(cv::Scalar src2) {
-    return operate_t<I, binary_operation_scalar<binary_sub<CUDA_T(I)>, CUDA_T(I), CUDA_T(I)>>()(src2);
+fk::binary_operation_scalar<fk::binary_sub<CUDA_T(I)>, CUDA_T(I), CUDA_T(I)> subtract(cv::Scalar src2) {
+    return internal::operate_t<I, fk::binary_operation_scalar<fk::binary_sub<CUDA_T(I)>, CUDA_T(I), CUDA_T(I)>>()(src2);
 }
 
 template <int I>
-binary_operation_scalar<binary_div<CUDA_T(I)>, CUDA_T(I), CUDA_T(I)> divide(cv::Scalar src2) {
-    return operate_t<I, binary_operation_scalar<binary_div<CUDA_T(I)>, CUDA_T(I), CUDA_T(I)>>()(src2);
+fk::binary_operation_scalar<fk::binary_div<CUDA_T(I)>, CUDA_T(I), CUDA_T(I)> divide(cv::Scalar src2) {
+    return internal::operate_t<I, fk::binary_operation_scalar<fk::binary_div<CUDA_T(I)>, CUDA_T(I), CUDA_T(I)>>()(src2);
 }
 
 template <int I>
-binary_operation_scalar<binary_sum<CUDA_T(I), CUDA_T(I)>, CUDA_T(I), CUDA_T(I)> add(cv::Scalar src2) {
-    return operate_t<I, binary_operation_scalar<binary_sum<CUDA_T(I), CUDA_T(I)>, CUDA_T(I), CUDA_T(I)>>()(src2);
+fk::binary_operation_scalar<fk::binary_sum<CUDA_T(I), CUDA_T(I)>, CUDA_T(I), CUDA_T(I)> add(cv::Scalar src2) {
+    return internal::operate_t<I, fk::binary_operation_scalar<fk::binary_sum<CUDA_T(I), CUDA_T(I)>, CUDA_T(I), CUDA_T(I)>>()(src2);
 }
 
 template <int I>
-split_write_scalar<perthread_split_write<CUDA_T(I)>, CUDA_T(I)> split(std::vector<cv::cuda::GpuMat>& output) {
-    return split_t<I, split_write_scalar<perthread_split_write<CUDA_T(I)>, CUDA_T(I)>>()(output);
+fk::split_write_scalar<fk::perthread_split_write<CUDA_T(I)>, CUDA_T(I)> split(std::vector<cv::cuda::GpuMat>& output) {
+    return internal::split_t<I, fk::split_write_scalar<fk::perthread_split_write<CUDA_T(I)>, CUDA_T(I)>>()(output);
 }
 
 template <int I, typename... operations>
@@ -66,7 +66,7 @@ void executeOperations(const cv::cuda::GpuMat& input, cv::cuda::Stream& stream, 
     dim3 grid(ceil(num_elems / (float)block.x));
     cudaStream_t cu_stream = cv::cuda::StreamAccessor::getStream(stream);
 
-    cuda_transform_noret<<<grid, block, 0, cu_stream>>>(num_elems, static_cast<CUDA_T(I)*>(static_cast<void*>(input.data)), ops...);
+    fk::cuda_transform_noret<<<grid, block, 0, cu_stream>>>(num_elems, static_cast<CUDA_T(I)*>(static_cast<void*>(input.data)), ops...);
     gpuErrchk(cudaGetLastError());
 }
 
@@ -78,9 +78,9 @@ void executeOperations(const cv::cuda::GpuMat& input, cv::cuda::GpuMat& output, 
     dim3 grid(ceil(num_elems / (float)block.x));
     cudaStream_t cu_stream = cv::cuda::StreamAccessor::getStream(stream);
 
-    memory_write_scalar<perthread_write<CUDA_T(O)>, CUDA_T(O), CUDA_T(O)> opFinal = { static_cast<CUDA_T(O)*>(static_cast<void*>(output.data)) };
+    fk::memory_write_scalar<fk::perthread_write<CUDA_T(O)>, CUDA_T(O), CUDA_T(O)> opFinal = { static_cast<CUDA_T(O)*>(static_cast<void*>(output.data)) };
 
-    cuda_transform_noret<<<grid, block, 0, cu_stream>>>(num_elems, static_cast<CUDA_T(I)*>(static_cast<void*>(input.data)), ops..., opFinal);
+    fk::cuda_transform_noret<<<grid, block, 0, cu_stream>>>(num_elems, static_cast<CUDA_T(I)*>(static_cast<void*>(input.data)), ops..., opFinal);
     gpuErrchk(cudaGetLastError());
 }
