+++ include/fast_kernel/cuda_vector_utils.h
@@ -19,10 +19,14 @@
 
 #define NUM_COMPONENTS(some_type) (sizeof(some_type)/sizeof(decltype(some_type::x)))
 
-// Automagically making any CUDA vector type from a template type
-// It will not compile if you try to do bad things. The number of elements
-// need to conform to T, and the type of the elements will always be casted.
-template <typename T, typename... Numbers>
-inline constexpr __device__ __host__ T make_(Numbers... pack) {
-    return {static_cast<decltype(T::x)>(pack)...};
+namespace fk {
+
+    // Automagically making any CUDA vector type from a template type
+    // It will not compile if you try to do bad things. The number of elements
+    // need to conform to T, and the type of the elements will always be casted.
+    template <typename T, typename... Numbers>
+    inline constexpr __device__ __host__ T make_(Numbers... pack) {
+        return {static_cast<decltype(T::x)>(pack)...};
+    }
+
 }
